Combine parse and config methods into main file
