+++ package.json
@@ -1,14 +1,12 @@
 {
   "name": "snackables",
-  "version": "0.0.12",
+  "version": "0.0.13",
   "description": "Loading byte sized ENVs made simple.",
   "main": "lib/index.js",
   "sideEffects": false,
   "homepage": "https://github.com/mattcarlotta/snackables",
   "files": [
-    "lib/config.js",
     "lib/index.js",
-    "lib/parse.js",
     "types"
   ],
   "types": "types/index.d.ts",
@@ -61,7 +59,7 @@
     "jest": "^26.6.0",
     "prettier": "^2.1.1",
     "rimraf": "^3.0.2",
-    "snackables": "^0.0.11",
+    "snackables": "^0.0.12",
     "terser": "^5.3.7",
     "ts-jest": "^26.3.0",
     "ts-node": "^9.0.0",
